♻️refactor #143: IconColor 타입을 공용 타입 Color로 변경
- 사용처에 적용
- designSystem에서 사용하지 않는 iconFilter 제거
